Add immediate validation on observations post (#109)
* use smallint for current quality flags

* add solarforecastarbiter to requirements

* adjust auth to include access token proxy

* move observation value validation, enqueue immediate validation

* update validate_observation_values

* fix tests that would require another api call

* include range in error

* move mocked_validation and avoid making actual requests

+++ datastore/migrations/0003_add_other_tables.up.sql
@@ -62,7 +62,7 @@ CREATE TABLE arbiter_data.observations_values (
     id BINARY(16) NOT NULL,
     timestamp TIMESTAMP NOT NULL,
     value FLOAT NOT NULL,
-    quality_flag TINYINT UNSIGNED NOT NULL,
+    quality_flag SMALLINT UNSIGNED NOT NULL,
 
     PRIMARY KEY (id, timestamp),
     FOREIGN KEY (id)
